% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_matrices.R
\docType{data}
\name{abundMats}
\alias{abundMats}
\title{Assemblage data from Calatayud et al.}
\format{A list of 385 assemblages, each represented by a site by species matrix with
\describe{
  \item{rows}{representing sites}
  \item{columns}{representing species}
  \item{cells}{recording the abundance or percent cover of each species in each site}
}}
\source{
\url{https://figshare.com/articles/Positive_associations_among_rare_species_and_their_persistence_in_ecological_assemblages/9906092}
}
\usage{
abundMats
}
\description{
A list of site by species matrices, each containing counts or percent
covers of different species observed across locations
}
\references{
Calatayud, J., et al. (2019). Positive associations among rare species and their persistence in ecological assemblages. Nature Ecology & Evolution, 1-6.
}
\keyword{datasets}
