% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdistrNB.R
\name{fitdistrNB}
\alias{fitdistrNB}
\title{Maximum likelihood of negative binomial distribution}
\usage{
fitdistrNB(x)
}
\arguments{
\item{x}{an integer vector}
}
\value{
A \code{lsit} with components \code{estiamte} giving the MLE and \code{loglik}
giving the value of the log likelihood function at the maximum. If
\code{var(x) < mean(x)} then the MLE does not exist and the funciton returns
\code{size = Inf}.
}
\description{
Finds the maximum likelihood estimates of the dispersion (k) and mean (mu)
parameters of the negative binomial distribution.
}
\details{
This function lets \code{mu = mean(x)} and uses \code{optimize} to find
\code{k} with log likelihood function defined simply as \code{sum(dnbinom(x, k, mu))}
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
