% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbFit.R
\name{nbFit}
\alias{nbFit}
\title{Fit the negative binomial}
\usage{
nbFit(x)
}
\arguments{
\item{x}{a site by species matrix}
}
\value{
A \code{matrix} with columns giving the species IDs, the two parameters of the
negative binomial distribution, the log likelihood of the negative binomial
distribution, and the log likelihood of the Poisson distribution.  Each row is one
species and the species ID refers to the column index of that species in the site x
species matrix \code{x}.
}
\description{
Fits the negative binomial SSAD to a site x species matrix
}
\details{
If likelihood optimization fails \code{NA} will be returned.
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
