% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbLLZ.R
\name{nbLLZ}
\alias{nbLLZ}
\title{Z-squared test statistic of the negative binomial distribution}
\usage{
nbLLZ(x, size, mu)
}
\arguments{
\item{x}{a numeric vector}

\item{size}{the size parameter}

\item{mu}{the mean parameter}
}
\value{
A \code{matrix} with columns giving the species IDs, the two parameters of the
negative binomial distribution, the log likelihood of the negative binomial
distribution, and the log likelihood of the Poisson distribution.  Each row is one
species and the species ID refers to the column index of that species in the site x
species matrix \code{x}.
}
\description{
Exact test of the negative binomial
}
\details{
Compares the observed log likelihood of the nagative binomial distribution
given the data to the sampling distribution of log likelihoods given the negative
binomial distribution is true.  This test statistic is Chi-square distributed with
d.f. = 1, meaning value larger than \code{qchisq(0.95, 1)} rejects the "null"
hypothesis that the data come from a negative binomial.
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
