% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusMinus.R
\name{plusMinus}
\alias{plusMinus}
\title{Positive and negative associations}
\usage{
plusMinus(x, alpha = 0.05, B = 999)
}
\arguments{
\item{x}{site by species matrix (sites as rows, species as columns, abundances in cells)}

\item{alpha}{the significance level}

\item{B}{number of replicates for the null model permutations}
}
\value{
an \code{ncol(x)} by \code{ncol(x)} matrix of species-species similarities
}
\description{
Calculate positive and negative associations based on Schoener similarity
}
\details{
Faster implementation of the funciton \code{pos.neg.abun.r2d} from ((citation))
that also only returns information relevant to testing the association of abundance and
different types of connectivity (plus and minus).
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
