% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPlusMinus.R
\name{simPlusMinus}
\alias{simPlusMinus}
\alias{simpleSim}
\title{Simulate positive and negative interactions from purely random data}
\usage{
simPlusMinus(nsite, nspp, mcCores, ssadType, nsim)

simpleSim(nsite, nspp, mcCores, sadfun, ssadfun, nsim)
}
\arguments{
\item{nsite}{number of sites to simulate}

\item{nspp}{number of species to simulate}

\item{mcCores}{number of cores to use in \code{parallel::mclapply}}

\item{ssadType}{string specifying SSAD shape (e.g. \code{'nbinom'})}

\item{nsim}{number of simulations to run}

\item{sadfun}{function to generate random SAD sample}

\item{ssadfun}{function to generate random SSAD sample}
}
\value{
an \code{ncol(x)} by \code{ncol(x)} matrix of species-species similarities
}
\description{
Simulate spatial replicates of abundance data and apply the same
analytical pipeline to those data that would be applied to real data
}
\details{
Note: any value passed to \code{ssadType} other than \code{'nbinom'} results
in a Poisson SSAD (i.e., there are only two options, negative binomial specified by
\code{'nbinom'} or Poisson specified by anything else)
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
