% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssadSim.R
\name{ssadSim}
\alias{ssadSim}
\title{Simulate permuted SSADs}
\usage{
ssadSim(nsite, nspp, mcCores, sadfun, ssadfun, nsim)
}
\arguments{
\item{nsite}{number of sites to simulate}

\item{nspp}{number of species to simulate}

\item{mcCores}{number of cores to use in \code{parallel::mclapply}}

\item{sadfun}{function to generate random SAD sample}

\item{ssadfun}{function to generate random SSAD sample}

\item{nsim}{number of simulations to run}
}
\value{
a \code{data.frame} with columns giving the z-score of the two negative
binomial parameters, and the difference in AIC values of the negative binomial versus
the Poisson
}
\description{
Simulate many SSADs and permute them to understand the effect of
permutation on the shape of the SSAD
}
\details{
Note, \code{sadfun} must be a function that takes a single argument \code{n}
(number of species), e.g., \code{function(n) pika::rfish(n, b = 0.1)}. Similarely,
\code{ssadfun} must take both \code{n} (the number of sites) and also \code{mu} the
mean abundance, e.g. \code{function(n) rnbinom(n, size = 0.1, mu = mu)}.
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
